Extract `PluginSwitcher` component so it can be reused for the mobile view
