+++ client/dashboard/plugins/manage/index.tsx
@@ -6,34 +6,26 @@ import {
 } from '@automattic/api-queries';
 import { useQuery } from '@tanstack/react-query';
 import { useParams } from '@tanstack/react-router';
-import {
-	__experimentalGrid as Grid,
-	__experimentalVStack as VStack,
-	Icon,
-} from '@wordpress/components';
+import { __experimentalGrid as Grid } from '@wordpress/components';
 import { filterSortAndPaginate, View } from '@wordpress/dataviews';
-import { __, _n, sprintf } from '@wordpress/i18n';
-import { plugins as pluginIcon } from '@wordpress/icons';
-import clsx from 'clsx';
+import { __ } from '@wordpress/i18n';
 import { useMemo } from 'react';
-import { pluginRoute } from '../../app/router/plugins';
 import { Card, CardBody } from '../../components/card';
 import { OptInWelcome } from '../../components/opt-in-welcome';
 import { PageHeader } from '../../components/page-header';
 import PageLayout from '../../components/page-layout';
 import { SectionHeader } from '../../components/section-header';
-import SwitcherContent from '../../components/switcher/switcher-content';
 import { Text } from '../../components/text';
 import { TextBlur } from '../../components/text-blur';
 import { PluginTabs } from '../plugin';
 import { usePlugin } from '../plugin/use-plugin';
+import { PluginSwitcher } from './components/plugin-switcher';
 import { useSitesById } from './hooks/use-sites-by-id';
 import { mapApiPluginsToDataViewPlugins } from './utils';
 import type { PluginListRow } from './types';
+
 import './style.scss';
 
-const ICON_SIZE = 40;
-const FALLBACK_ICON_SIZE = 30;
 const view: View = {
 	type: 'list',
 	page: 1,
@@ -154,78 +146,12 @@ export default function PluginsList() {
 			notices={ <OptInWelcome tracksContext="plugins" /> }
 		>
 			<Grid columns={ 2 } gap={ 6 } templateColumns="392px 1fr">
-				<Card>
-					<CardBody className="plugin-switcher-card-body">
-						<SwitcherContent
-							itemClassName={ ( item ) =>
-								clsx( 'plugin-switcher-item', { 'is-selected': selectedPluginSlug === item.slug } )
-							}
-							searchClassName="plugin-switcher-search"
-							initialView={ view }
-							items={ pluginsWithIcon }
-							resetScroll={ false }
-							getItemUrl={ ( item ) => pluginRoute.to.replace( '$pluginId', item.slug ) }
-							renderItemMedia={ ( { item } ) => {
-								const icon = item.icon ? (
-									<img
-										src={ item.icon }
-										alt={ item.name }
-										width={ ICON_SIZE }
-										height={ ICON_SIZE }
-									/>
-								) : (
-									<Icon
-										icon={ pluginIcon }
-										size={ FALLBACK_ICON_SIZE }
-										className="plugin-icon-fallback"
-									/>
-								);
-
-								return (
-									<div className={ clsx( 'plugin-icon-wrapper', { 'is-fallback': ! item.icon } ) }>
-										{ icon }
-									</div>
-								);
-							} }
-							renderItemTitle={ ( { item } ) => {
-								const sitesText = sprintf(
-									// translators: %(siteCount)d is the number of sites the plugin is installed on.
-									_n( '%(siteCount)d site', '%(siteCount)d sites', item.sitesCount ),
-									{ siteCount: item.sitesCount }
-								);
-
-								const updatesText = item.sitesWithPluginUpdate.length
-									? sprintf(
-											// translators: %(updateCount)d is the number of updates available.
-											_n(
-												'%(updateCount)d update available',
-												'%(updateCount)d updates available',
-												item.sitesWithPluginUpdate.length
-											),
-											{ updateCount: item.sitesWithPluginUpdate.length }
-									  )
-									: '';
-
-								return (
-									<VStack spacing={ 0 }>
-										{ /* @ts-expect-error: Can only set one of `children` or `props.dangerouslySetInnerHTML`. */ }
-										<Text
-											className="plugin-switcher-item-name"
-											dangerouslySetInnerHTML={ { __html: item.name } }
-											title={ item.name }
-										/>
-										<Text variant="muted">
-											{ updatesText ? `${ sitesText }, ${ updatesText }` : sitesText }
-										</Text>
-									</VStack>
-								);
-							} }
-							searchableFields={ searchableFields }
-							onClose={ () => {} }
-							width="auto"
-						/>
-					</CardBody>
-				</Card>
+				<PluginSwitcher
+					pluginsWithIcon={ pluginsWithIcon }
+					searchableFields={ searchableFields }
+					selectedPluginSlug={ selectedPluginSlug }
+					view={ view }
+				/>
 
 				<Card>
 					<CardBody className="plugin-sites-card-body">
