+++ client/dashboard/plugins/manage/style.scss
@@ -1,16 +1,3 @@
-.plugin-icon-wrapper {
-	border-radius: 4px;
-	position: relative;
-	width: 40px;
-	height: 40px;
-	flex-shrink: 0;
-
-	&.is-fallback {
-		background: #c3c4c7;
-		fill: #fff;
-	}
-}
-
 .plugin-icon-placeholder {
 	width: 32px;
 	height: 32px;
@@ -25,54 +12,13 @@ img.plugin-icon {
 	height: 32px;
 }
 
-.plugin-icon-fallback {
-	position: absolute;
-	top: 50%;
-	inset-inline-start: 50%;
-	transform: translate(-50%, -50%);
-}
-
-.plugin-switcher-card-body,.plugin-sites-card-body {
+.plugin-sites-card-body {
 	// take into account header and nav plus border (64+1), page layout padding (32),
 	// page title/description (80), and gap between title/description and the cards (32)
 	max-height: calc(100vh - 2 * ( 64px + 1px ) - 2 * 32px - 80px - 32px);
 	padding-inline-start: 0;
 	padding-inline-end: 0;
-}
-
-.plugin-switcher-card-body {
-	overflow-y: auto;
-}
-
-.plugin-sites-card-body {
 	overflow-y: hidden;
-}
-
-.plugin-switcher-search {
-	padding: 0 calc(4px * 6) calc(4px * 4);
-	width: fit-content;
-}
-
-.plugin-switcher-item {
-	border-top: 1px solid #f0f0f0;
-	padding: calc(4px * 2.5) calc(4px * 6);
-
-	&.is-selected,
-	&:hover {
-		background-color: #f8f8f8;
-	}
-
-	&:focus:not(:disabled),
-	&:focus-visible:not(:disabled) {
-		box-shadow: none;
-		outline: none;
-	}
-}
-
-.plugin-switcher-item-name {
-	overflow: hidden;
-	text-overflow: ellipsis;
-	text-wrap-mode: nowrap;
 }
 
 .plugin-sites-card-header {
